Move version guessing in `VersionGuesser` class
